+++ pf1.js
@@ -842,10 +842,11 @@ Hooks.on("dropActorSheetData", (act, sheet, data) => {
 Hooks.on("hotbarDrop", (bar, data, slot) => {
   let macro;
   if (data.type === "Item") macro = macros.createItemMacro(data.data, slot);
-  if (data.type === "skill") macro = macros.createSkillMacro(data.skill, data.actor, slot);
-  if (data.type === "save") macro = macros.createSaveMacro(data.altType, data.actor, slot);
-  if (["defenses", "cmb", "concentration", "cl", "bab"].includes(data.type))
+  else if (data.type === "skill") macro = macros.createSkillMacro(data.skill, data.actor, slot);
+  else if (data.type === "save") macro = macros.createSaveMacro(data.altType, data.actor, slot);
+  else if (["defenses", "cmb", "concentration", "cl", "bab"].includes(data.type))
     macro = macros.createMiscActorMacro(data.type, data.actor, slot, data.altType);
+  else return true;
 
   if (macro === undefined) return false;
 });